# packages/types/openapi.json
{
  "components": {
    "schemas": {
...
      "CellOut": {
        "properties": {
          "accession": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accession"
          },
          "concept": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept"
          },
          "derivation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derivation"
          },
          "filed_date": {
            "format": "date",
            "title": "Filed Date",
            "type": "string"
          },
          "period_key": {
            "title": "Period Key",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "period_key",
          "value",
          "concept",
          "accession",
          "filed_date",
          "derivation"
        ],
        "title": "CellOut",
        "type": "object"
      },
...
      "CompanySearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompanySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "CompanySearchResponse",
        "type": "object"
      },
      "CompanySummary": {
        "properties": {
          "cik": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cik"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "name",
          "exchange",
          "cik",
          "sector",
          "industry",
          "country",
          "is_active"
        ],
        "title": "CompanySummary",
        "type": "object"
      },
...
      "FundamentalsResponse": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "cik": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cik"
          },
          "formula_version": {
            "title": "Formula Version",
            "type": "string"
          },
          "mapping_version": {
            "title": "Mapping Version",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricSeriesOut"
            },
            "title": "Metrics",
            "type": "array"
          },
          "period_type": {
            "enum": [
              "annual",
              "quarterly"
            ],
            "title": "Period Type",
            "type": "string"
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/PeriodOut"
            },
            "title": "Periods",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            },
            "title": "Sources",
            "type": "array"
          },
          "statements": {
            "$ref": "#/components/schemas/StatementsOut"
          },
          "status": {
            "enum": [
              "current",
              "stale",
              "unavailable",
              "not_configured",
              "not_applicable",
              "not_available"
            ],
            "title": "Status",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "cik",
          "period_type",
          "as_of",
          "status",
          "message",
          "mapping_version",
          "formula_version",
          "periods",
          "statements",
          "metrics",
          "sources"
        ],
        "title": "FundamentalsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LineItemOut": {
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/CellOut"
            },
            "title": "Cells",
            "type": "array"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "unit": {
            "enum": [
              "USD",
              "USD/shares",
              "shares"
            ],
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "unit",
          "cells"
        ],
        "title": "LineItemOut",
        "type": "object"
      },
...
      "MetricPointOut": {
        "properties": {
          "period_key": {
            "title": "Period Key",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "period_key",
          "value"
        ],
        "title": "MetricPointOut",
        "type": "object"
      },
      "MetricSeriesOut": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "formula": {
            "title": "Formula",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "unit": {
            "enum": [
              "percent",
              "ratio",
              "currency",
              "shares",
              "per_share"
            ],
            "title": "Unit",
            "type": "string"
          },
          "values": {
            "items": {
              "$ref": "#/components/schemas/MetricPointOut"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "key",
          "label",
          "category",
          "unit",
          "formula",
          "values"
        ],
        "title": "MetricSeriesOut",
        "type": "object"
      },
...
      "PeerRow": {
        "properties": {
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "is_subject": {
            "title": "Is Subject",
            "type": "boolean"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Metrics",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sic_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sic Code"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "name",
          "sic_code",
          "industry",
          "is_subject",
          "metrics"
        ],
        "title": "PeerRow",
        "type": "object"
      },
      "PeersResponse": {
        "properties": {
          "basis": {
            "title": "Basis",
            "type": "string"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PeerRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "basis",
          "columns",
          "rows"
        ],
        "title": "PeersResponse",
        "type": "object"
      },
      "PeriodOut": {
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "fiscal_quarter": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiscal Quarter"
          },
          "fiscal_year": {
            "title": "Fiscal Year",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "fiscal_year",
          "fiscal_quarter",
          "start",
          "end"
        ],
        "title": "PeriodOut",
        "type": "object"
      },
...
      "SourceRef": {
        "description": "Where a set of values came from. Every externally sourced field links to one.",
        "properties": {
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "fetch_id": {
            "title": "Fetch Id",
            "type": "integer"
          },
          "license_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Note"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "fetch_id",
          "provider",
          "dataset",
          "source_url",
          "retrieved_at",
          "license_note"
        ],
        "title": "SourceRef",
        "type": "object"
      },
      "StatementsOut": {
        "properties": {
          "balance": {
            "items": {
              "$ref": "#/components/schemas/LineItemOut"
            },
            "title": "Balance",
            "type": "array"
          },
          "cash_flow": {
            "items": {
              "$ref": "#/components/schemas/LineItemOut"
            },
            "title": "Cash Flow",
            "type": "array"
          },
          "income": {
            "items": {
              "$ref": "#/components/schemas/LineItemOut"
            },
            "title": "Income",
            "type": "array"
          }
        },
        "required": [
          "income",
          "balance",
          "cash_flow"
        ],
        "title": "StatementsOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Pinero Research API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/companies": {
      "get": {
        "operationId": "search_api_v1_companies_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanySearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "companies"
        ]
      }
    },
...
    "/api/v1/companies/{ticker}/fundamentals": {
      "get": {
        "operationId": "fundamentals_api_v1_companies__ticker__fundamentals_get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "annual",
              "enum": [
                "annual",
                "quarterly"
              ],
              "title": "Period",
              "type": "string"
            }
          },
          {
            "description": "Only data public on or before",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only data public on or before",
              "title": "As Of"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 40,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundamentalsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fundamentals",
        "tags": [
          "fundamentals"
        ]
      }
    },
...
    "/api/v1/companies/{ticker}/peers": {
      "get": {
        "operationId": "peers_api_v1_companies__ticker__peers_get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated extra peers",
            "in": "query",
            "name": "tickers",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated extra peers",
              "title": "Tickers"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 25,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Peers",
        "tags": [
          "fundamentals"
        ]
      }
    },
...
  }
}